+++ src/geometry/triangle.cpp
@@ -1,7 +1,7 @@
 #include "geometry/triangle.h"
 
 
-bool Triangle::containsEdge(int v1, int v2) const {
+bool Triangle::containsEdge(size_t v1, size_t v2) const {
     return (v[0] == v1 && v[1] == v2) || (v[1] == v1 && v[0] == v2) ||
             (v[1] == v1 && v[2] == v2) || (v[2] == v1 && v[1] == v2) ||
             (v[2] == v1 && v[0] == v2) || (v[0] == v1 && v[2] == v2);
